refactor visualization
